services/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
presidio-analyzer>=2.2.352
presidio-anonymizer>=2.2.352
pydantic>=1.10.0
python-multipart>=0.0.5
socksio>=1.0.0
litellm>=1.45.0
spacy-langdetect>=0.1.2
langdetect>=1.0.9
